requirements.txt
PyPDF2==3.0.1
pydantic==2.4.2
google-generative-ai==0.3.1 